Run a quick test for the Gravatar URL
